Refactor css styling in it's own file

+++ src/main/java/com/rossie/demo/model/Person.java
@@ -1,4 +1,4 @@
 package com.rossie.demo.model;
 
-public record Person(int id, String firstName, String lastName, int age, String email) {
+public record Person(long id, String firstName, String lastName, int age, String email) {
 }
